--- a/pkg/jsonnet/testdata/findImporters/environments/chain-env1/config.jsonnet
+++ b/pkg/jsonnet/testdata/findImporters/environments/chain-env1/config.jsonnet
@@ -1,0 +1,4 @@
+{
+  database: "postgres://env1-db/data",
+  feature_flag: "enabled"
+}