--- a/pkg/jsonnet/testdata/findImporters/environments/chain-env1/main.jsonnet
+++ b/pkg/jsonnet/testdata/findImporters/environments/chain-env1/main.jsonnet
@@ -1,0 +1,4 @@
+{
+  name: "env1",
+  config: "original-config"
+}