--- a/pkg/jsonnet/testdata/findImporters/environments/chain-env2/main.jsonnet
+++ b/pkg/jsonnet/testdata/findImporters/environments/chain-env2/main.jsonnet
@@ -1,0 +1,9 @@
+local lib1 = import 'chain-lib1/main.libsonnet';
+
+{
+  // env2 imports lib1
+  name: "env2",
+  inherited_config: lib1.processedConfig + {
+    enhanced_by: "env2"
+  }
+}