--- a/pkg/jsonnet/testdata/findImporters/environments/lib-imports-environment/config.jsonnet
+++ b/pkg/jsonnet/testdata/findImporters/environments/lib-imports-environment/config.jsonnet
@@ -1,0 +1,4 @@
+{
+  shared_config: "environment-specific-value",
+  database_url: "postgres://localhost/test"
+}