--- a/pkg/jsonnet/testdata/findImporters/environments/lib-imports-environment/main.jsonnet
+++ b/pkg/jsonnet/testdata/findImporters/environments/lib-imports-environment/main.jsonnet
@@ -1,0 +1,6 @@
+{
+  environment: "test-environment",
+  config: {
+    value: 42
+  }
+}