--- a/pkg/jsonnet/testdata/findImporters/environments/uses-lib-that-imports-env/main.jsonnet
+++ b/pkg/jsonnet/testdata/findImporters/environments/uses-lib-that-imports-env/main.jsonnet
@@ -1,0 +1,6 @@
+local lib = import 'lib-imports-environment/main.libsonnet';
+
+{
+  // This environment uses a lib that imports from another environment
+  result: lib.environment_config
+}