--- a/pkg/jsonnet/testdata/findImporters/lib/chain-lib1/main.libsonnet
+++ b/pkg/jsonnet/testdata/findImporters/lib/chain-lib1/main.libsonnet
@@ -1,0 +1,6 @@
+{
+  // lib1 imports env1
+  processedConfig: (import '../../environments/chain-env1/config.jsonnet') + {
+    processed_by: "lib1"
+  }
+}