--- a/pkg/jsonnet/testdata/findImporters/lib/chain-lib3/main.libsonnet
+++ b/pkg/jsonnet/testdata/findImporters/lib/chain-lib3/main.libsonnet
@@ -1,0 +1,6 @@
+{
+  // lib3 imports env2
+  importedFromEnv2: (import '../../environments/chain-env2/main.jsonnet') + {
+    processed_by: "lib3"
+  }
+}