--- a/pkg/jsonnet/testdata/findImporters/lib/lib-imports-environment/main.libsonnet
+++ b/pkg/jsonnet/testdata/findImporters/lib/lib-imports-environment/main.libsonnet
@@ -1,0 +1,8 @@
+{
+  // This lib file imports from an environment directory. So when we look for
+  // the importers of `config.jsonnet` we should find environments that import
+  // this file - those are the ones that need to be re-exported.
+  environment_config: (import '../../environments/lib-imports-environment/config.jsonnet') + {
+    lib_enhancement: 'added by lib',
+  },
+}