@iainlane 2025-09-17
I found there's a benchmark suite! I backported the new tests to `main`, hacked the benchmark to not be able to fail (IMO it shouldn't, it should just run the code). Here are the results:

```
goos: linux
goarch: amd64
pkg: github.com/grafana/tanka/pkg/jsonnet
cpu: AMD Ryzen 7 PRO 6860Z with Radeon Graphics
                                                                                                     │  before.txt  │              after.txt              │
                                                                                                     │    sec/op    │    sec/op     vs base               │
FindImporters-16                                                                                       14.55m ± 10%   14.82m ± 14%       ~ (p=0.529 n=10)
FindImporters_StaticCases/no_files-16                                                                  8.391µ ± 12%   8.882µ ± 13%       ~ (p=0.631 n=10)
FindImporters_StaticCases/invalid_file-16                                                              51.39µ ± 14%   48.59µ ± 12%       ~ (p=0.089 n=10)
FindImporters_StaticCases/project_with_no_imports-16                                                   874.7µ ± 12%   819.8µ ±  5%       ~ (p=0.315 n=10)
FindImporters_StaticCases/local_import-16                                                              820.0µ ±  5%   797.0µ ±  7%       ~ (p=0.165 n=10)
FindImporters_StaticCases/local_import_with_relative_path-16                                           879.2µ ± 16%   843.6µ ± 10%       ~ (p=0.353 n=10)
FindImporters_StaticCases/lib_imported_through_chain-16                                                834.2µ ±  9%   877.9µ ±  6%       ~ (p=0.075 n=10)
FindImporters_StaticCases/vendored_lib_imported_through_chain_+_directly-16                            1.720m ± 10%   1.671m ±  7%       ~ (p=0.436 n=10)
FindImporters_StaticCases/vendored_lib_found_through_symlink-16                                        884.1µ ±  7%   858.6µ ± 15%       ~ (p=0.796 n=10)
FindImporters_StaticCases/text_file-16                                                                 819.0µ ± 11%   833.0µ ±  7%       ~ (p=0.971 n=10)
FindImporters_StaticCases/relative_imported_environment-16                                             831.4µ ± 10%   798.4µ ± 11%       ~ (p=0.631 n=10)
FindImporters_StaticCases/relative_imported_environment_with_doubled_'..'-16                           884.0µ ± 10%   900.7µ ± 13%       ~ (p=0.529 n=10)
FindImporters_StaticCases/relative_imported_text_file-16                                               816.4µ ±  8%   801.8µ ±  7%       ~ (p=0.353 n=10)
FindImporters_StaticCases/relative_imported_text_file_with_doubled_'..'-16                             875.8µ ± 11%   811.7µ ±  5%       ~ (p=0.105 n=10)
FindImporters_StaticCases/vendor_override_in_env:_override_vendor_used-16                              855.7µ ±  9%   823.9µ ±  8%       ~ (p=0.529 n=10)
FindImporters_StaticCases/vendor_override_in_env:_global_vendor_unused-16                              831.2µ ±  7%   838.7µ ±  6%       ~ (p=0.481 n=10)
FindImporters_StaticCases/imported_file_in_lib_relative_to_env-16                                      849.8µ ±  7%   817.3µ ±  8%       ~ (p=0.063 n=10)
FindImporters_StaticCases/unused_deleted_file-16                                                       22.19µ ± 10%   22.56µ ± 15%       ~ (p=0.971 n=10)
FindImporters_StaticCases/deleted_local_path_that_is_still_potentially_imported-16                     24.19µ ± 11%   24.85µ ±  9%       ~ (p=0.631 n=10)
FindImporters_StaticCases/deleted_lib_that_is_still_potentially_imported-16                            23.87µ ± 13%   22.64µ ± 10%       ~ (p=0.739 n=10)
FindImporters_StaticCases/deleted_vendor_that_is_still_potentially_imported-16                         22.38µ ± 10%   23.85µ ±  8%       ~ (p=0.247 n=10)
FindImporters_StaticCases/deleted_lib_that_is_still_potentially_imported,_relative_path_from_root-16   20.70µ ± 14%   22.37µ ± 26%       ~ (p=0.143 n=10)
FindImporters_StaticCases/deleted_dir_in_environment-16                                                23.38µ ±  7%   23.87µ ±  7%       ~ (p=0.280 n=10)
FindImporters_StaticCases/imports_through_a_main_file_are_followed-16                                  832.1µ ±  7%   812.4µ ±  8%       ~ (p=0.631 n=10)
FindImporters_StaticCases/lib_file_imports_environment_file-16                                         875.8µ ±  9%   824.9µ ± 11%       ~ (p=0.143 n=10)
FindImporters_StaticCases/complex_transitive_chain:_env1_->_lib1_->_env2_->_lib3_->_env3-16            829.5µ ± 10%   861.5µ ±  8%       ~ (p=0.247 n=10)
geomean                                                                                                317.3µ         315.1µ        -0.70%

                                                                                                     │  before.txt  │               after.txt               │
                                                                                                     │     B/op     │     B/op      vs base                 │
FindImporters-16                                                                                       1.632Mi ± 0%   1.631Mi ± 0%       ~ (p=0.218 n=10)
FindImporters_StaticCases/no_files-16                                                                    834.0 ± 0%     834.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/invalid_file-16                                                              4.641Ki ± 0%   4.641Ki ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/project_with_no_imports-16                                                   47.77Ki ± 0%   47.76Ki ± 0%       ~ (p=0.271 n=10)
FindImporters_StaticCases/local_import-16                                                              47.83Ki ± 0%   47.83Ki ± 0%       ~ (p=0.698 n=10)
FindImporters_StaticCases/local_import_with_relative_path-16                                           47.84Ki ± 0%   47.83Ki ± 0%       ~ (p=0.493 n=10)
FindImporters_StaticCases/lib_imported_through_chain-16                                                47.83Ki ± 0%   47.86Ki ± 0%       ~ (p=0.271 n=10)
FindImporters_StaticCases/vendored_lib_imported_through_chain_+_directly-16                            97.29Ki ± 0%   97.26Ki ± 0%       ~ (p=0.812 n=10)
FindImporters_StaticCases/vendored_lib_found_through_symlink-16                                        48.73Ki ± 0%   48.74Ki ± 0%       ~ (p=0.425 n=10)
FindImporters_StaticCases/text_file-16                                                                 49.10Ki ± 0%   49.11Ki ± 0%       ~ (p=0.138 n=10)
FindImporters_StaticCases/relative_imported_environment-16                                             47.83Ki ± 0%   47.82Ki ± 0%       ~ (p=0.060 n=10)
FindImporters_StaticCases/relative_imported_environment_with_doubled_'..'-16                           47.83Ki ± 0%   47.85Ki ± 0%       ~ (p=0.127 n=10)
FindImporters_StaticCases/relative_imported_text_file-16                                               47.79Ki ± 0%   47.77Ki ± 0%       ~ (p=0.109 n=10)
FindImporters_StaticCases/relative_imported_text_file_with_doubled_'..'-16                             47.78Ki ± 0%   47.76Ki ± 0%  -0.04% (p=0.045 n=10)
FindImporters_StaticCases/vendor_override_in_env:_override_vendor_used-16                              47.88Ki ± 0%   47.91Ki ± 0%       ~ (p=0.066 n=10)
FindImporters_StaticCases/vendor_override_in_env:_global_vendor_unused-16                              47.75Ki ± 0%   47.75Ki ± 0%       ~ (p=0.782 n=10)
FindImporters_StaticCases/imported_file_in_lib_relative_to_env-16                                      47.86Ki ± 0%   47.87Ki ± 0%       ~ (p=0.190 n=10)
FindImporters_StaticCases/unused_deleted_file-16                                                       1.988Ki ± 0%   1.988Ki ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/deleted_local_path_that_is_still_potentially_imported-16                     2.160Ki ± 0%   2.160Ki ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/deleted_lib_that_is_still_potentially_imported-16                            2.066Ki ± 0%   2.066Ki ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/deleted_vendor_that_is_still_potentially_imported-16                         2.066Ki ± 0%   2.066Ki ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/deleted_lib_that_is_still_potentially_imported,_relative_path_from_root-16   1.973Ki ± 0%   1.973Ki ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/deleted_dir_in_environment-16                                                2.160Ki ± 0%   2.160Ki ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/imports_through_a_main_file_are_followed-16                                  47.94Ki ± 0%   47.93Ki ± 0%       ~ (p=0.128 n=10)
FindImporters_StaticCases/lib_file_imports_environment_file-16                                         47.83Ki ± 0%   47.93Ki ± 0%  +0.21% (p=0.000 n=10)
FindImporters_StaticCases/complex_transitive_chain:_env1_->_lib1_->_env2_->_lib3_->_env3-16            47.81Ki ± 0%   48.09Ki ± 0%  +0.58% (p=0.000 n=10)
geomean                                                                                                21.37Ki        21.38Ki       +0.03%
¹ all samples are equal

                                                                                                     │ before.txt  │              after.txt               │
                                                                                                     │  allocs/op  │  allocs/op   vs base                 │
FindImporters-16                                                                                       22.50k ± 0%   22.50k ± 0%       ~ (p=1.000 n=10)
FindImporters_StaticCases/no_files-16                                                                   9.000 ± 0%    9.000 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/invalid_file-16                                                               57.00 ± 0%    57.00 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/project_with_no_imports-16                                                    560.0 ± 0%    560.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/local_import-16                                                               561.0 ± 0%    561.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/local_import_with_relative_path-16                                            561.0 ± 0%    561.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/lib_imported_through_chain-16                                                 562.0 ± 0%    562.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/vendored_lib_imported_through_chain_+_directly-16                            1.135k ± 0%   1.135k ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/vendored_lib_found_through_symlink-16                                         570.0 ± 0%    570.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/text_file-16                                                                  573.0 ± 0%    573.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/relative_imported_environment-16                                              562.0 ± 0%    562.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/relative_imported_environment_with_doubled_'..'-16                            562.0 ± 0%    562.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/relative_imported_text_file-16                                                561.0 ± 0%    561.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/relative_imported_text_file_with_doubled_'..'-16                              561.0 ± 0%    561.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/vendor_override_in_env:_override_vendor_used-16                               561.0 ± 0%    561.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/vendor_override_in_env:_global_vendor_unused-16                               559.0 ± 0%    559.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/imported_file_in_lib_relative_to_env-16                                       561.0 ± 0%    561.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/unused_deleted_file-16                                                        23.00 ± 0%    23.00 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/deleted_local_path_that_is_still_potentially_imported-16                      25.00 ± 0%    25.00 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/deleted_lib_that_is_still_potentially_imported-16                             25.00 ± 0%    25.00 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/deleted_vendor_that_is_still_potentially_imported-16                          25.00 ± 0%    25.00 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/deleted_lib_that_is_still_potentially_imported,_relative_path_from_root-16    25.00 ± 0%    25.00 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/deleted_dir_in_environment-16                                                 25.00 ± 0%    25.00 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/imports_through_a_main_file_are_followed-16                                   563.0 ± 0%    563.0 ± 0%       ~ (p=1.000 n=10) ¹
FindImporters_StaticCases/lib_file_imports_environment_file-16                                          561.0 ± 0%    563.0 ± 0%  +0.36% (p=0.000 n=10)
FindImporters_StaticCases/complex_transitive_chain:_env1_->_lib1_->_env2_->_lib3_->_env3-16             561.0 ± 0%    565.0 ± 0%  +0.71% (p=0.000 n=10)
geomean                                                                                                 252.7         252.8       +0.04%
¹ all samples are equal
```

So there's essentially no difference.